combine spu/sc command extensions in the cluster extension
