+++ src/cli/Cargo.toml
@@ -17,23 +17,7 @@ path = "src/bin/main.rs"
 doc = false
 
 [features]
-default = ["cluster_components", "native2_tls"]
-cluster_components = [
-    "k8-client/native_tls",
-    "fluvio-spu",
-    "fluvio-sc/k8",
-    "fluvio/native_tls",
-    "fluvio-cluster/native_tls",
-    "openssl/vendored",
-]
-cluster_components_rustls = [
-    "k8-client/rust_tls",
-    "fluvio-spu",
-    "fluvio-sc/k8_rustls",
-    "fluvio/rust_tls",
-    "fluvio-cluster/rust_tls",
-    "openssl/vendored",
-]
+default = ["native2_tls"]
 native2_tls = ["fluvio-future/native2_tls", "fluvio-extension-consumer/native_tls"]
 
 [build-dependencies]
@@ -72,16 +56,7 @@ openssl = { version = "0.10.30", optional = true }
 
 flv-util = { version = "0.5.0" }
 fluvio-future = { version = "0.1.8", features = ["fs", "io", "subscriber"] }
-k8-client = { version = "3.0.0", default-features = false }
-k8-config = { version = "1.3.0", features = ["context"] }
-k8-obj-core = { version = "1.1.0" }
-k8-obj-metadata = { version = "1.0.0" }
-k8-metadata-client = { version = "1.0.0" }
 fluvio = { version = "0.2.3", path = "../client", default-features = false }
-fluvio-sc = { version = "0.3.1", path = "../sc", optional = true, default-features = false }
-fluvio-sc-schema = { version = "0.2.0", path = "../sc-schema", features = ["use_serde"] }
-fluvio-spu = { version = "0.2.0", path = "../spu", optional = true }
-fluvio-controlplane-metadata = { version = "0.2.0", path = "../controlplane-metadata", features = ["use_serde", "k8"] }
 fluvio-cluster = { version = "0.3.1", path = "../cluster", default-features = false, features = ["cmd_extension"] }
 fluvio-types = { path = "../types", version = "0.1.0" }
 fluvio-package-index = { version = "0.2.0", path = "../package-index" }
