+++ src/cluster/Cargo.toml
@@ -14,8 +14,8 @@ path = "src/lib.rs"
 
 [features]
 default = ["native_tls"]
-rust_tls = ["fluvio/rust_tls","k8-client/rust_tls","fluvio-runner/rust_tls"]
-native_tls = ["fluvio/native_tls","k8-client/native_tls","fluvio-runner/native_tls"]
+rust_tls = ["fluvio/rust_tls","k8-client/rust_tls"]
+native_tls = ["fluvio/native_tls","k8-client/native_tls"]
 cmd_extension = ["structopt","fluvio-extension-common","prettytable-rs"]
 
 [dependencies]
@@ -42,7 +42,7 @@ k8-client = { version = "3.0.0", default-features = false }
 k8-obj-core = "1.1.0"
 k8-obj-metadata = "1.0.0"
 fluvio-extension-common = { version = "0.1.0", path = "../extension-common", optional = true }
-fluvio-runner = { versio = "0.1.0", path = "../extension-runner"}
+fluvio-runner = { version = "0.1.0", path = "../extension-runner"}
 
 [dev-dependencies]
 fluvio-future = { version = "0.1.0", features =["task"]}
