+++ src/cluster/src/extension/install/local.rs
