+++ src/cluster/src/extension/uninstall.rs
